pyproject.toml
[project]
name = "dynamiccomfy"
version = "0.1.0"
description = "MVP orchestration service for ComfyUI docker instances"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0",
  "docker>=7.1.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "passlib>=1.7.4",
  "pydantic-settings>=2.5.2",
  "pyjwt>=2.9.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.35",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
]

[tool.uv]
package = false
